Fix time zone issue for event start and end times.
In the database start and end times have a timezone.  However, OGO
appears to convert these timestamps to UTC and interprets the result
as local time.  We output the calendar in UTC so we have to take
account for this disparity. This is done by:

1. Obtaining the timestamp in UTC;
2. Interpreting this result as Local time;
3. Converting this Local timestamp into UTC.

Assuming Local time is UTC+1 this will effectively produce a timestamp
that is two hours behind the original Local database
timestamp. Assuming Local time is UTC+0 this will result in a
timestamp that equals the database timestamp.  This is accomplished
with the following SQL:

((((timestamp AT TIME ZONE 'UTC')::timestamp without time zone)::timestamp with time zone) AT TIME ZONE 'UTC')