Ensure that dates account for timezone differences.
Datetimes coming from the database are in UTC.  When converted
directly to a date UTC dates don't match the original input date if
applying the original timezone shifts the date backwards or forwards.
This fixes the issue but assumes that the server is running in the
same timezone as the OGO database.